requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
streamlit==1.36.0
beautifulsoup4==4.12.3
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
spacy==3.7.4
langchain==0.2.6
langchain-google-genai==1.0.6
sqlalchemy==2.0.31
python-dotenv==1.0.1
plotly==5.22.0
